Refactor: Create MicroagentManagementClient interface and extract microagent methods from BaseGitService
- Created new MicroagentManagementClient interface in integrations/protocols/
- Moved all microagent-related abstract and concrete methods from BaseGitService to the new interface
- Updated BaseGitService to be a proper abstract class (inherits from ABC) without microagent functionality
- Updated all provider service feature mixins (GitHub, GitLab, BitBucket) to implement MicroagentManagementClient
- Fixed import paths for MicroagentParseError to use the new interface location
- All provider services now inherit microagent functionality through their feature mixins rather than BaseGitService

This separation of concerns improves code organization and makes microagent management capabilities explicit.
