+++ .openhands/TASKS.md
@@ -1,0 +1,13 @@
+# Task List
+
+1. ✅ Explore the current structure of integrations/protocols and BaseGitService
+
+2. ✅ Create MicroagentManagementClient interface in integrations/protocols
+
+3. ✅ Identify and extract microagent-related methods from BaseGitService
+
+4. ✅ Update all provider services to implement MicroagentManagementClient
+
+5. ✅ Verify all implementations are correct and consistent
+
+6. ✅ Ensure BaseGitService is properly abstract (inherits from ABC) without MicroagentManagementClient
