+++ openhands/integrations/bitbucket/service/features.py
@@ -1,10 +1,13 @@
 from openhands.core.logger import openhands_logger as logger
 from openhands.integrations.bitbucket.service.base import BitBucketMixinBase
+from openhands.integrations.protocols.microagent_management_client import (
+    MicroagentManagementClient,
+)
 from openhands.integrations.service_types import ResourceNotFoundError
 from openhands.microagent.types import MicroagentContentResponse
 
 
-class BitBucketFeaturesMixin(BitBucketMixinBase):
+class BitBucketFeaturesMixin(BitBucketMixinBase, MicroagentManagementClient):
     """
     Mixin for BitBucket feature operations (microagents, cursor rules, etc.)
     """
